third_party/WebKit/LayoutTests/http/tests/preload/preload-csp.html

<!DOCTYPE html>
<meta http-equiv="Content-Security-Policy" content="font-src 'none'">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
    var t = async_test('Makes sure that preload requests respect CSP');
</script>
<link rel=preload href="../resources/Ahem.ttf" as=font crossorigin>
<script src="../resources/slow-script.pl?delay=200"></script>
<script>
    window.onload = t.step(function(){
        var entries = performance.getEntriesByName(new URL("../resources/Ahem.ttf", location.href).href);
        assert_equals(entries.length, 0);
        t.done();
    });
</script>
